refactor(projet): broken down the client page
