+++ components/projects/index.ts
@@ -1,0 +1,6 @@
+// components/projects/form/index.ts
+export { default as ProjectCreationForm } from "@/components/projects/form/sections/ProjectCreationForm";
+export { default as ProjectInformationSection } from "@/components/projects/form/sections/ProjectInformationSection";
+export { default as ProjectSettingsSection } from "@/components/projects/form/sections/ProjectSettingsSection";
+export { default as ProjectFormNavigation } from "@/components/projects/form/NavigationSidebar";
+export { default as ProjectAlertMessages } from "@/components/projects/form/AlertMessages";
