+++ types/index.ts
@@ -88,13 +88,22 @@ export type {
   BaseNavSource,
 } from "./utility";
 
-// Non-object types
+// Teams
 export type {
   InviteFormMember,
   PendingTeamMember,
-  FormData,
+  TeamFormData,
   TeamSettings,
 } from "./forms/team";
+
+export type {
+  ProjectSettings,
+  ProjectCreationProps,
+  ProjectFormSectionProps,
+  ProjectFormNavigationProps,
+  ProjectFormMessagesProps,
+  ProjectFormData,
+} from "./forms/project";
 
 // Notification
 export type {
